Clean up: FaceStyle and GlyphImageHeightRatio.
* Face.ParseLegacyStyle now returns FaceStyle, whose values are compatible with
  System.Drawing.FontStyle. This change is meant to make clear the link between the two.
* Replace the hard-coded value 0.95 with GlyphImageHeightRatio, which represents the
  ratio between a glyph and its rendered image.

+++ KiriProj/Project.cs
@@ -93,7 +93,7 @@ namespace KiriProj
             using ( var face = lib.OpenFace( spec.FontPath, spec.FaceIndex, openParams ) )
             {
                 projectFile.FontFamily = face.FamilyName;
-                projectFile.FontStyle = Face.ParseLegacyStyle( face.StyleName );
+                projectFile.FontStyle = (int) Face.ParseLegacyStyle( face.StyleName );
             }
         }
